backend/requirements.txt
# FastAPI and server
fastapi>=0.128.0,<0.129.0
uvicorn[standard]>=0.30.0,<0.31.0
python-multipart>=0.0.9,<0.1.0

# Database
sqlalchemy>=2.0.45,<2.1.0
aiosqlite>=0.20.0,<0.21.0

# Validation
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.2.0,<3.0.0

# Background tasks
celery>=5.6.1,<5.7.0
redis>=5.2.0,<6.0.0

# AI - Anthropic
anthropic>=0.40.0,<0.41.0

# AI - LangChain
langchain>=0.3.0,<0.4.0
langchain-anthropic>=0.2.0,<0.3.0
langgraph>=0.2.0,<0.3.0

# Code analysis (tree-sitter-languages optional, doesn't support Python 3.12+)
# tree-sitter>=0.25.0,<0.26.0
# tree-sitter-languages>=1.10.0,<2.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
httpx>=0.27.0,<0.28.0

# GitHub App
PyJWT>=2.8.0,<3.0.0
cryptography>=42.0.0,<43.0.0
